Wire up file size

+++ Toolkit.Foundation/FileDescriptor.cs
@@ -1,4 +1,4 @@
 ﻿namespace Toolkit.Foundation;
 
-public record FileDescriptor(string Name, string Path, int Size) :
+public record FileDescriptor(string Name, string Path, long Size) :
     IFileDescriptor;
